refactor(v1.13.15): split surface_manifest.rs into runtime vs docs/export
Per architecture audit P2.1 (2026-04-24): "Split surface_manifest.rs
into runtime manifest vs docs/export adapters." Runtime concerns —
server card, language inventory, profile/preset enumeration — stay in
the parent module. The static contract surfaces consumed by external
docs and resource URIs move to a sibling sub-module.

Moved to surface_manifest/exports.rs:
- build_agent_experience() (171 LOC of static contract data)
- handoff_artifact_schema_json() (pub(crate), reachable from resources.rs)
- build_harness_artifacts_summary()
- 5 schema_version / doc_path / resource_uri constants
- HANDOFF_ARTIFACT_SCHEMA_TEXT (compile-time include_str! body)

LOC distribution:
- surface_manifest.rs: 609 → 406 (-203, runtime only)
- surface_manifest/exports.rs: new 220 (docs/export only)

resources.rs path preserved via `pub(crate) use exports::handoff_artifact_schema_json`
re-export from the parent module — no breakage for external callers
that resolve `crate::surface_manifest::handoff_artifact_schema_json`.

Pure refactor, zero behavior change. Tests:
- cargo build --release --workspace: passed (59s)
- cargo test -p codelens-mcp --bin: 536 passed, 0 failed
- surface_manifest::tests: 3 passed

Follow-up (v1.13.16+ candidate): tighten visibility of the 4 export
constants from pub(crate) to pub(super) — the audit moved them but
left visibility wider than necessary.
